Re-added road and background sprites and created a game scene for Antony
- placed the road and background sprites back into the scene and rescalled them to look more in proportion
- changed one line of code in the road spawner script so i can change the roadsize width as a float
- re positioned the camera to give a better overall layout
- created a scene for anotny to work on the interface inclding a reference image to get an idea of the space he can work with.

+++ Double Spawner/Assets/My Scripts/GameScene/RoadSpawner.cs
@@ -8,7 +8,9 @@ public class RoadSpawner : MonoBehaviour {
     public GameObject roadSquare2;
     public GameObject roadSquare3;
 
-    public int roadSquareWidth = 10;
+    public float roadSquareWidth = 10.0f;
+    //*Gordon* - changed roadSquarewidth from 'int' to 'float'
+
     public GameObject roadLocationSpawn;
 
     private int WhichRoadType;
